requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
pymongo==4.5.0
aiohttp==3.8.5
